For feed title, subtitle, etc.: run the clean-html hooks when the values are added to the database, rather than when the strings are pulled out to be printed to the user, for speed.
Fix the HTMLPurifier plugin to accept a maximum string length: if the
input string is "<div>some string", HTMLPurifier will append a
"</div>" to make it balance. The maximum length means it'll try to
truncate the original string to "<div>some " to make it all work out.
This because these strings must fit inside a finite-sized field in the
database.